refactor(streaming): extract useFrameRenderer hook
The canvas-render pipeline — bitmap decode queue + scheduling cursor +
rAF draw loop, plus the canvas ref registration — moves into its own
hook. ~110 lines and four refs (`bitmapQueue`, `lastScheduledAt`,
`frameTimeline`, `drawRaf`) plus the `canvasReady` state are no longer
visible at provider level.

The hook takes the per-frame inputs (`frame`, the three frame-metric
refs) as args and returns just what the provider re-exposes:
`registerCanvas`, `canvasReady`, `frameTimelineRef`. Render-pipeline
internals stay scoped to the file.
